third_party/WebKit/LayoutTests/http/tests/shapedetection/shapedetection-cross-origin.html

<!DOCTYPE html>
<script src=../../../../resources/testharness.js></script>
<script src=../../../../resources/testharnessreport.js></script>
<script>

// http:// resources are considered cross-origin from the current file://.
const IMAGE_URL = "http://localhost:8080/security/resources/abe.png";

// TODO(xianglu): Move this helper function to a shared location for reuse.
function detectFaceOnImageElementAndExpectError(imageUrl) {
  return new Promise(function(resolve, reject) {
    var image = new Image();
    var tryFaceDetection = function() {
      var faceDetector = new FaceDetector();
      faceDetector.detect(image)
          .then(faceDetectionResult => {
            reject("Promise should have been rejected.");
          })
          .catch(error => {
            resolve(error);
          });
    };
    image.onload = tryFaceDetection;
    image.onerror = tryFaceDetection;
    image.src = imageUrl;
  });
}

function detectFaceOnImageBitmapAndExpectError(imageUrl) {
  return new Promise(function(resolve, reject) {
    var image = new Image();
    image.onload = function() {
      createImageBitmap(image)
          .then(imageBitmap => {
            var faceDetector = new FaceDetector();
            return faceDetector.detect(imageBitmap);
          })
          .then(faceDetectionResult => {
            reject("Promise should have been rejected.");
          })
          .catch(error => {
            resolve(error);
          });
    };
    image.onerror = () => {};  // Explicitly ignore expected error events.
    image.src = imageUrl;
  });
}

// Verifies that FaceDetector rejects a cross-origin HTMLImageElement.
promise_test(function(t) {
  return detectFaceOnImageElementAndExpectError(IMAGE_URL)
      .then(error => {
        assert_equals(error.name, "SecurityError");
      });
},
"FaceDetector should reject cross-origin HTMLImageElements with a SecurityError.");

// Verifies that FaceDetector rejects a cross-origin ImageBitmap.
promise_test(function(t) {
  return detectFaceOnImageBitmapAndExpectError(IMAGE_URL)
      .then(error => {
        assert_equals(error.name, "SecurityError");
      });
},
"FaceDetector should reject cross-origin ImageBitmaps with a SecurityError.");

</script>
